test(blockchain-link): co-locate tests
