+++ packages/blockchain-link/jest.config.integration.cjs
@@ -7,5 +7,13 @@ const baseConfig = require('../../jest.config.base.swc');
 module.exports = {
     ...baseConfig,
     moduleFileExtensions: ['js', 'ts', 'json'],
-    testMatch: ['**/tests/integration/*.test.ts'],
+    testMatch: [
+        '<rootDir>/src/workers/blockbook/blockbook.test.ts',
+        '<rootDir>/src/workers/blockfrost/blockfrost.test.ts',
+        '<rootDir>/src/workers/connection.test.ts',
+        '<rootDir>/src/workers/electrum/electrum.test.ts',
+        '<rootDir>/src/workers/ripple/ripple.test.ts',
+        '<rootDir>/src/workers/solana/solana.test.ts',
+        '<rootDir>/src/workers/stellar/stellar.test.ts',
+    ],
 };
