+++ packages/blockchain-link/src/__fixtures__/getAccountInfo-ripple.ts
