+++ packages/blockchain-link/src/connection.test.ts
@@ -4,8 +4,9 @@ import { BackendWebsocketServerMock } from '@trezor/e2e-utils';
 // eslint-disable-next-line @typescript-eslint/no-restricted-imports
 import { TimeoutError } from '@trezor/network-ripple';
 
-import workers from './worker';
-import { BlockchainLink } from '../../src';
+import workers from './__fixtures__/worker';
+
+import { BlockchainLink } from './index';
 
 const getMethod = (instanceName: string) => {
     let method: string;
