+++ packages/blockchain-link/src/estimateFee.test.ts
@@ -1,11 +1,12 @@
 import { BackendWebsocketServerMock } from '@trezor/e2e-utils';
 
-import workers from './worker';
-import { BlockchainLink } from '../../src';
-import fixtures from './fixtures/getBlockHash';
+import fixtures from './__fixtures__/estimateFee';
+import workers from './__fixtures__/worker';
+
+import { BlockchainLink } from './index';
 
 workers.forEach(instance => {
-    describe(`getBlockHash: ${instance.name}`, () => {
+    describe(`estimateFee: ${instance.name}`, () => {
         let server: BackendWebsocketServerMock;
         let blockchain: BlockchainLink;
 
@@ -23,13 +24,15 @@ workers.forEach(instance => {
             blockchain.dispose();
             await server.close();
         };
+
         beforeAll(setup);
         afterAll(teardown);
 
         fixtures[instance.name].forEach(f => {
             it(f.description, async () => {
                 server.setFixtures(f.serverFixtures);
-                const promise = blockchain.getBlockHash(f.params);
+                // @ts-expect-error
+                const promise = blockchain.estimateFee(f.params);
                 if (!f.error) {
                     expect(await promise).toEqual(f.response);
                 } else {
