+++ packages/blockchain-link/src/getBlockHash.test.ts
@@ -1,11 +1,12 @@
 import { BackendWebsocketServerMock } from '@trezor/e2e-utils';
 
-import workers from './worker';
-import { BlockchainLink } from '../../src';
-import fixtures from './fixtures/getAccountInfo';
+import fixtures from './__fixtures__/getBlockHash';
+import workers from './__fixtures__/worker';
+
+import { BlockchainLink } from './index';
 
 workers.forEach(instance => {
-    describe(`getAccountInfo: ${instance.name}`, () => {
+    describe(`getBlockHash: ${instance.name}`, () => {
         let server: BackendWebsocketServerMock;
         let blockchain: BlockchainLink;
 
@@ -26,21 +27,10 @@ workers.forEach(instance => {
         beforeAll(setup);
         afterAll(teardown);
 
-        // [btc-unknown-tx-debug] getAccountInfo parses tx history through transformTransaction, which
-        // emits a temporary console.error for txs classified as 'unknown' with account context. Silence
-        // the JestCustomEnv console.error trap for these fixtures.
-        beforeEach(() => {
-            jest.spyOn(console, 'error').mockImplementation(() => {});
-        });
-        afterEach(() => {
-            jest.restoreAllMocks();
-        });
-
         fixtures[instance.name].forEach(f => {
             it(f.description, async () => {
                 server.setFixtures(f.serverFixtures);
-                // @ts-expect-error incorrect params
-                const promise = blockchain.getAccountInfo(f.params);
+                const promise = blockchain.getBlockHash(f.params);
                 if (!f.error) {
                     expect(await promise).toEqual(f.response);
                 } else {
