+++ packages/blockchain-link/src/workers/solana/solana.test.ts
@@ -1,8 +1,9 @@
 import type { AccountInfoParams } from '@trezor/blockchain-link-types';
 import type { SolanaAPI } from '@trezor/network-solana/types';
 
-import { BlockchainLink } from '../../src';
-import SolanaWorker from '../../src/workers/solana';
+import { BlockchainLink } from '../../index';
+
+import SolanaWorker from './index';
 
 const id = 79;
 const descriptor = '2MLmmoKgCrxVEzMeGatnjdABYS5RXsQSNikcWrmnvQna';
