+++ packages/blockchain-link/src/workers/utils.test.ts
@@ -1,5 +1,5 @@
-import * as fixtures from './fixtures/utils';
-import { prioritizeEndpoints } from '../../src/workers/utils';
+import * as fixtures from './__fixtures__/utils';
+import { prioritizeEndpoints } from './utils';
 
 describe('prioritizeEndpoints', () => {
     it('prioritizeEndpoints', () => {
